Создание связи и использование проекта smpl-security

+++ src/main/java/com/api/smpl_library_api/service/AuthorService.java
@@ -1,9 +1,11 @@
 package com.api.smpl_library_api.service;
 
+import com.api.smpl_library_api.dto.UserDTO;
 import com.api.smpl_library_api.model.Author;
 import com.api.smpl_library_api.repository.AuthorRepository;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,14 +16,22 @@ import java.util.Optional;
 public class AuthorService {
     private final AuthorRepository authorRepository;
 
+    private final SecurityConnector securityConnector;
+
     @Transactional
-    public Author save(Author author) {
-        return authorRepository.save(author);
+    public Optional<Author> save(Author author, UserDTO userDTO) {
+        if (securityConnector.checkUser(userDTO)) {
+            return Optional.of(authorRepository.save(author));
+        }
+        return Optional.empty();
     }
 
     @Transactional
-    public boolean update(Author author) {
-        return authorRepository.update(author) > 0L;
+    public boolean update(Author author, UserDTO userDTO) {
+        if (securityConnector.checkUser(userDTO)) {
+            return authorRepository.update(author) > 0L;
+        }
+        return false;
     }
 
     public Optional<Author> findById(Integer id) {
@@ -29,8 +39,11 @@ public class AuthorService {
     }
 
     @Transactional
-    public boolean deleteById(Integer id) {
-        return authorRepository.delete(id) > 0L;
+    public boolean deleteById(Integer id, UserDTO userDTO) {
+        if (securityConnector.checkUser(userDTO)) {
+            return authorRepository.delete(id) > 0L;
+        }
+        return false;
     }
 
     public List<Author> findAll() {
